Refactor PaletteContainer into submodule container/components
